requirements.txt
Flask==3.0.0
Pillow==10.1.0
pytesseract==0.3.10
gunicorn==21.2.0
pytest==7.4.3